Add create profile, move funcs out of constants
